refactor: sidebar 컴포넌트명 변경 및 about 페이지 구조 변경
