+++ src/app/about/_components/side-content.tsx
@@ -1,0 +1,3 @@
+const SideContent = () => <div className='flex-1 p-6'>우측 컨텐츠</div>;
+
+export default SideContent;
